requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
PyPDF2==3.0.1
pdfplumber==0.10.3
nltk==3.8.1
spacy==3.7.4
gunicorn==21.2.0
python-dotenv==1.0.1
Flask-WTF==1.2.1
SQLAlchemy==2.0.25
pytest==8.0.2
python-magic==0.4.27
Pillow==9.5.0 